- Removed unused feature packages - Renamed stable package to riders
