--- app/src/main/java/com/example/stablestable/ui/calendar/CalendarScreen.kt
@@ -1,15 +1,0 @@
-package com.example.stablestable.ui.calendar
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.navigation.NavController
-
-@Composable
-fun CalendarScreen(
-    navController: NavController
-) {
-    Column {
-        Text(text = "Calendar / Weekly Schedule Screen")
-    }
-}
