+++ app/src/main/java/com/example/stablestable/ui/riders/StableScreen.kt
@@ -1,13 +1,11 @@
-package com.example.stablestable.ui.stable
+package com.example.stablestable.ui.riders
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.navigation.NavController
 
 @Composable
 fun StableScreen(
-    navController: NavController
 ) {
     Column {
         Text(text = "Stable Information Screen")
