+++ app/src/main/java/com/example/stablestable/ui/riders/UserList.kt
@@ -1,4 +1,4 @@
-package com.example.stablestable.components
+package com.example.stablestable.ui.riders
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.lazy.LazyColumn
@@ -10,15 +10,12 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import androidx.navigation.NavController
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.stablestable.navigation.AuthViewModel
-import com.example.stablestable.ui.profile.ProfileViewModel
-import com.example.stablestable.ui.stable.StableUsersViewModel
 
 @Composable
 fun StableUsers(
-    navController: NavController,
+
 ) {
     val authViewModel: AuthViewModel = viewModel()
     val stableUsersViewModel: StableUsersViewModel = viewModel(factory = object : ViewModelProvider.Factory {
